Finalised json to bcf conversion.
