+++ README.md
@@ -3,20 +3,24 @@
 A .NET library and a command line utility for converting `BCF` (Building 
 Collaboration Format) files into `json` and vice versa.
 
-The tool converts `BCF` information across formats and versions. 
+The tool converts `BCF` information across formats ~~and versions~~. 
 
 ## Usage
 
 ### CLI
 
 The command line interface accepts three arguments:
- * the source file
- * the target file
+ * the source bcf file or json folder
+ * the target bcf file or json folder
  * (optional) the BCF version of the output, the default is 2.1
- * (the version of the source is inferred from the file)
+ 
+The json representation is one file for every Markup, while the BCF format
+is a zipped file as per the standard.
 
 ```
-$ ifc-converter /path/to/source.bcfzip /path/to/target.json 2.1
+$ ifc-converter /path/to/source.bcfzip /path/to/target/folder 2.1
+
+$ ifc-converter /path/to/json/folder /path/to/target/bcf.bcfzip 2.1
 ```
 
 ## Development
@@ -26,12 +30,17 @@ completed features.
 
 |          | XML 2.0 | XML 2.1 | JSON 2.0 | JSON 2.1 |
 |----------|:-------:|:-------:|:--------:|:--------:|
-| XML 2.0  |         |         |          |     X    |
-| XML 2.1  |         |         |          |          |
+| XML 2.0  |         |         |          |          |
+| XML 2.1  |         |         |          |     X    |
 | JSON 2.0 |         |         |          |          |
-| JSON 2.1 |         |         |          |          |
+| JSON 2.1 |         |    X    |          |          |
 
+The models for the BCF in-memory representation were auto-created from the
+[latest XSDs][1] using the [`XmlSchemaClassGenerator`][2].
 
 ### Contribution
 
 Code style guide can be found in the `bcf-converter.sln.DotSettings` file.
+
+[1]: https://github.com/buildingSMART/BCF-XML/tree/master/Schemas
+[2]: https://github.com/mganss/XmlSchemaClassGenerator
