+++ bcf-converter/Parser/Xml21/XElementExtensions.cs
@@ -2,7 +2,7 @@ using System;
 using System.Xml.Linq;
 using bcf_converter.Model;
 
-namespace bcf_converter.Parser.Xml20 {
+namespace bcf_converter.Parser.Xml21 {
   /// <summary>
   ///   A list of convenience extension methods on the XElementExtensions class.
   ///   The BCF version specific parsing logic is defined here.
